// src/types/Ticket.tsx
{/*  ===================================THIS FILE WAS AUTO GENERATED=================================== */}

export type Ticket = {
    ticketID: number;
    title: string;
    description: string;
    status: string;
    createdBy: string;
    createdAt: string;
    dueDate: string;
    completedAt: string;
    userID: number;
};